requirements.txt
numpy==1.26.4
opencv-python
torch
Pillow
torchvision
mss
pywin32
